+++ src/jit_solver/components/agk47.cpp
@@ -3,14 +3,14 @@
 #include "../state.h"
 
 template <typename Provider>
-void AGK47<Provider>::execute(SimulationState& st, float /*dt*/) {
+void AGK47<Provider>::execute(SimulationState& st, double /*dt*/) {
     // Push model: single-port component (connects to ground)
     // Just reads the input value - no output to write in push model
     (void)st.values[provider.get(PortNames::input)];
 }
 
 template <typename Provider>
-void AGK47<Provider>::commit(SimulationState& st, float /*dt*/) {
+void AGK47<Provider>::commit(SimulationState& st, double /*dt*/) {
     (void)st;
 }
