+++ src/jit_solver/components/and_gate.cpp
@@ -2,7 +2,7 @@
 #include "port_registry.h"
 
 template <typename Provider>
-void AND<Provider>::execute(SimulationState& st, float /*dt*/) {
+void AND<Provider>::execute(SimulationState& st, double /*dt*/) {
     float A = st.values[provider.get(PortNames::A)];
     float B = st.values[provider.get(PortNames::B)];
     // Treat > 0.5V as TRUE, else FALSE
@@ -13,7 +13,7 @@ void AND<Provider>::execute(SimulationState& st, float /*dt*/) {
 }
 
 template <typename Provider>
-void AND<Provider>::commit(SimulationState& st, float /*dt*/) {
+void AND<Provider>::commit(SimulationState& st, double /*dt*/) {
     (void)st;
 }
