+++ src/jit_solver/components/blueprint_input.cpp
@@ -3,14 +3,14 @@
 #include "../state.h"
 
 template <typename Provider>
-void BlueprintInput<Provider>::execute(SimulationState& st, float /*dt*/) {
+void BlueprintInput<Provider>::execute(SimulationState& st, double /*dt*/) {
     // No-op - pass-through component (like Bus)
     // Union-find will collapse port to connected signal
     (void)st;
 }
 
 template <typename Provider>
-void BlueprintInput<Provider>::commit(SimulationState& st, float /*dt*/) {
+void BlueprintInput<Provider>::commit(SimulationState& st, double /*dt*/) {
     (void)st;
 }
