+++ src/jit_solver/components/blueprint_output.cpp
@@ -3,14 +3,14 @@
 #include "../state.h"
 
 template <typename Provider>
-void BlueprintOutput<Provider>::execute(SimulationState& st, float /*dt*/) {
+void BlueprintOutput<Provider>::execute(SimulationState& st, double /*dt*/) {
     // No-op - pass-through component (like Bus)
     // Union-find will collapse port to connected signal
     (void)st;
 }
 
 template <typename Provider>
-void BlueprintOutput<Provider>::commit(SimulationState& st, float /*dt*/) {
+void BlueprintOutput<Provider>::commit(SimulationState& st, double /*dt*/) {
     (void)st;
 }
