+++ src/jit_solver/components/comparator.h
@@ -17,7 +17,7 @@ public:
 
     Comparator() = default;
 
-    void execute(SimulationState& st, float dt);
-    void commit(SimulationState& st, float dt);
+    void execute(SimulationState& st, double dt);
+    void commit(SimulationState& st, double dt);
     void pre_load();
 };
