+++ src/jit_solver/components/divide.cpp
@@ -3,7 +3,7 @@
 #include <cmath>
 
 template <typename Provider>
-void Divide<Provider>::execute(SimulationState& st, float /*dt*/) {
+void Divide<Provider>::execute(SimulationState& st, double /*dt*/) {
     float A = st.values[provider.get(PortNames::A)];
     float B = st.values[provider.get(PortNames::B)];
     // Avoid division by zero - if B is effectively zero, output zero
@@ -12,7 +12,7 @@ void Divide<Provider>::execute(SimulationState& st, float /*dt*/) {
 }
 
 template <typename Provider>
-void Divide<Provider>::commit(SimulationState& st, float /*dt*/) {
+void Divide<Provider>::commit(SimulationState& st, double /*dt*/) {
     (void)st;
 }
