+++ src/jit_solver/components/electrical_conductance.cpp
@@ -3,14 +3,14 @@
 #include "../state.h"
 
 template <typename Provider>
-void ElectricalConductance<Provider>::execute(SimulationState& /*st*/, float /*dt*/) {
+void ElectricalConductance<Provider>::execute(SimulationState& /*st*/, double /*dt*/) {
     // No-op: ElectricalConductance is a solver-owned primitive.
     // Electrical propagation runs via the conductance matrix in the
     // electrical subsolver, not via push scheduler.
 }
 
 template <typename Provider>
-void ElectricalConductance<Provider>::commit(SimulationState& /*st*/, float /*dt*/) {
+void ElectricalConductance<Provider>::commit(SimulationState& /*st*/, double /*dt*/) {
     // No state to update.
 }
