+++ src/jit_solver/components/fuel_tank.h
@@ -14,15 +14,15 @@ public:
 
     Provider provider;
     float capacity = 1000.0f;   // Tank capacity (liters)
-    float level = 1000.0f;      // Current fuel amount (liters, state)
-    float next_level = 1000.0f; // Next-frame fuel amount (liters)
+    double level = 1000.0;      // Current fuel amount (liters, state)
+    double next_level = 1000.0; // Next-frame fuel amount (liters)
     float density = 0.78f;      // Fuel density (kg/L, kerosene TS-1)
     float consumption_rate = 0.0f; // Constant fuel draw (liters/second)
     float inv_capacity = 0.001f; // Precomputed 1/capacity
 
     FuelTank() = default;
 
-    void execute(SimulationState& st, float dt);
-    void commit(SimulationState& st, float dt);
+    void execute(SimulationState& st, double dt);
+    void commit(SimulationState& st, double dt);
     void pre_load();
 };
