+++ src/jit_solver/components/generator.h
@@ -19,7 +19,7 @@ public:
 
     Generator() = default;
 
-    void execute(SimulationState& st, float dt);
-    void commit(SimulationState& st, float dt);
+    void execute(SimulationState& st, double dt);
+    void commit(SimulationState& st, double dt);
     void pre_load();
 };
