+++ src/jit_solver/components/greater_eq.cpp
@@ -2,14 +2,14 @@
 #include "port_registry.h"
 
 template <typename Provider>
-void GreaterEq<Provider>::execute(SimulationState& st, float /*dt*/) {
+void GreaterEq<Provider>::execute(SimulationState& st, double /*dt*/) {
     float A = st.values[provider.get(PortNames::A)];
     float B = st.values[provider.get(PortNames::B)];
     st.values[provider.get(PortNames::o)] = (A >= B) ? 1.0f : 0.0f;
 }
 
 template <typename Provider>
-void GreaterEq<Provider>::commit(SimulationState& st, float /*dt*/) {
+void GreaterEq<Provider>::commit(SimulationState& st, double /*dt*/) {
     (void)st;
 }
