+++ src/jit_solver/components/high_power_load.cpp
@@ -4,7 +4,7 @@
 #include <cmath>
 
 template <typename Provider>
-void HighPowerLoad<Provider>::execute(SimulationState& st, float /*dt*/) {
+void HighPowerLoad<Provider>::execute(SimulationState& st, double /*dt*/) {
     // TODO(Phase 3): Implement power-balance model.
     // Currently a no-op stub — power_draw and min_voltage_diff are unused.
     // Push model needs current injection or voltage drop to model load behavior.
@@ -12,7 +12,7 @@ void HighPowerLoad<Provider>::execute(SimulationState& st, float /*dt*/) {
 }
 
 template <typename Provider>
-void HighPowerLoad<Provider>::commit(SimulationState& st, float /*dt*/) {
+void HighPowerLoad<Provider>::commit(SimulationState& st, double /*dt*/) {
     (void)st;
 }
