+++ src/jit_solver/components/radiator.h
@@ -15,7 +15,7 @@ public:
 
     Radiator() = default;
 
-    void execute(SimulationState& st, float dt);
-    void commit(SimulationState& st, float dt);
+    void execute(SimulationState& st, double dt);
+    void commit(SimulationState& st, double dt);
     void pre_load() {}
 };
